replay module split replay_main.cpp into separate compilation units
