+++ Tools/uorb_graph/create.py
@@ -240,7 +240,7 @@ class Graph(object):
 
 
         special_cases_pub = [
-    ('replay', r'replay_main\.cpp$', None, r'^sub\.orb_meta$'),
+    ('replay', r'Replay\.cpp$', None, r'^sub\.orb_meta$'),
     ('fw_pos_control_l1', r'FixedwingPositionControl\.cpp$', r'\b_attitude_setpoint_id=([^,)]+)', r'^_attitude_setpoint_id$'),
 
     ('mc_pos_control', r'mc_pos_control_main\.cpp$', r'\b_attitude_setpoint_id=([^,)]+)', r'^_attitude_setpoint_id$'),
