+++ src/modules/replay/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2016 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2016-2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -35,6 +35,10 @@ px4_add_module(
 	MAIN replay
 	COMPILE_FLAGS
 	SRCS
+		definitions.hpp
 		replay_main.cpp
-	DEPENDS
+		Replay.cpp
+		Replay.hpp
+		ReplayEkf2.cpp
+		ReplayEkf2.hpp
 	)
